+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [1.1.0]
+
+### Added
+
+* `fastmath.clustering` namespace
+
+### Changed
+
+* removed `k-means` from `fastmath.stats`
+
 ## [1.0.3]
 
 ### Fixed